#572 refactor(zigbee): Move the metrics
