+++ controllers/zigbee/zigbee_controller/sensor/zigbee_energy_monitor.py
@@ -1,4 +1,3 @@
-from enum import StrEnum, unique
 from typing import Dict
 
 from powerpi_common.logger import Logger
@@ -9,26 +8,9 @@ from zigpy.zcl.clusters import Cluster
 from zigpy.zcl.foundation import Attribute
 
 from zigbee_controller.device import ZigbeeController
+from zigbee_controller.sensor.metrics import Metric, MetricValue
 from zigbee_controller.zigbee import ZigbeeMixin
 from zigbee_controller.zigbee.mixins import ZigbeeReportMixin
-
-
-@unique
-class Metric(StrEnum):
-    POWER = 'power'
-    CURRENT = 'current'
-    VOLTAGE = 'voltage'
-
-
-@unique
-class MetricValue(StrEnum):
-    NONE = 'none'
-    READ = 'read'
-    VISIBLE = 'visible'
-
-    @staticmethod
-    def is_enabled(value):
-        return value in [MetricValue.READ, MetricValue.VISIBLE]
 
 
 class ZigbeeEnergyMonitorSensor(Sensor, ZigbeeReportMixin, ZigbeeMixin):
